Revert "Group 45 api to add a comment"
