+++ Backend/README.md
@@ -1,81 +1,24 @@
-## Project Name
+# ProjectName
 Agricom
 
 ## Overview
-Welcome to Agricom, an innovative agricultural platform designed to empower farmers by creating a centralized hub for sharing, exchanging, and enhancing crop-yielding techniques and knowledge. Our primary goal is to facilitate collaboration and knowledge sharing among farmers, enabling them to maximize yields, especially for specialized crops in their region, and effectively combat crop diseases.
+Our project is AgriCom. AgriCom is an innovative agricultural platform designed as a centralized hub for farmers to share, exchange, and enhance crop-yielding techniques and knowledge. The primary purpose of this software is to facilitate collaboration and knowledge sharing among farmers, enabling them to maximize yields, particularly for specialized crops in their region, and effectively address crop diseases.
 
 ## Installation
-To get started with Agricom, follow these installation steps:
-
-Install the required dependencies by running the following commands:
-npm install
-npm install --save-dev jest supertest
-npm install jest supertest mongoose
-
+npm i
 
 ## Usage
-User Authentication
-Agricom provides user authentication for a secure and personalized experience.
+1. The user can sign up for the application by giving the user info.The endpoint is `HTTP://127.0.0.0:3000/signup('user123','user123@gmail.com', 'password123')`.
+2. The user can log in to the application using the username and password. The endpoint is `HTTP://127.0.0.0:3000/login('user123@gmail.com', 'password123')`.
+3. The user can change the password if forgotten by using the forgotten password. The endpoint is `HTTP://127.0.0.0:3000/forgotten password('user123@gmail.com')`.
+# Example code snippet
+app.post('/login', async(req, res) => {
+    # Your login logic here
+    }
 
-1.Sign Up: Users can create an account by providing their information. Use the following endpoint to sign up:
-
-POST 'http://127.0.0.1:3000/signup'
-
-Example:
-{
-    "username": "user123",
-    "email": "user123@gmail.com",
-    "password": "password123"
-}
-
-2.Log In: Registered users can log in to their accounts using their email and password. Use the following endpoint for login:
-
-POST 'http://127.0.0.1:3000/login'
-
-Example:
-
-{
-    "email": "user123@gmail.com",
-    "password": "password123"
-}
-
-Forgot Password: In case a user forgets their password, they can reset it by providing their email address. Use the following endpoint for resetting the password:
-
-POST 'http://127.0.0.1:3000/forgotten-password'
-Example:
-
-{
-    "email": "user123@gmail.com"
-}
-
-Agricom offers several API endpoints to support user interactions. Here are some example code snippets:
-
-Post a User
-app.post('/user', async (req, res) => {
-    // Add your user registration logic here
-});
-
-Retrieve Posts
-app.get('/posts', async (req, res) => {
-    // Logic to retrieve posts
-});
-
-Add Comment to a Post
-app.post('/comment', async (req, res) => {
-    // Logic to add a comment to a post
-});
-
-Get Comments for a Post
-app.get('/comments/:postId', async (req, res) => {
-    // Logic to retrieve comments for a specific post
-});
-
+# Usage example
+`HTTP://127.0.0.0:3000/login('user123', 'password123')`
 
 ## Features
-Agricom offers the following key features:
-1.User authentication and management, including signup, login, and password reset functionality.
-2.Backend integration with a database for storing user and agricultural data.
-3.API endpoints to support user interactions with posts and comments.
-
-Feel free to explore Agricom and enhance your agricultural knowledge and practices. 
-Happy farming with Agricom! 🌱🚜
+  - Implement the backend flow for login, signup, and ForgetPassword.
+  - Implement backend database (DB) integration.
