+++ Backend/package.json
@@ -3,7 +3,7 @@
   "version": "1.0.0",
   "main": "index.js",
   "scripts": {
-    "test": "jest"
+    "test": "echo \"Error: no test specified\" && exit 1"
   },
   "keywords": [],
   "author": "",
@@ -15,18 +15,13 @@
     "express-session": "^1.17.3",
     "firebase-admin": "^11.11.0",
     "mongodb": "^6.1.0",
+    "mongoose": "^7.6.0",
     "nodemailer": "^6.9.5",
     "passport": "^0.6.0",
     "passport-local": "^1.0.0"
   },
   "description": "",
   "devDependencies": {
-    "@types/nodemailer": "^6.4.11",
-    "chai": "^4.3.10",
-    "jest": "^29.7.0",
-    "mocha": "^10.2.0",
-    "mongodb-memory-server": "^9.0.1",
-    "mongoose": "^7.6.3",
-    "supertest": "^6.3.3"
+    "@types/nodemailer": "^6.4.11"
   }
 }
