+++ Frontend/angular.json
@@ -100,8 +100,5 @@
         }
       }
     }
-  },
-  "cli": {
-    "analytics": "2ee6890b-a89d-40fd-95f2-e1e949a8f7d2"
   }
 }
