--- Frontend/src/app/feed/feed.component.html
@@ -1,35 +1,0 @@
-<div id="post-bar">
-    <div class="dropdown-container">
-        <button class="btn btn-primary" (click)="toggleDropdown()" id="dropdown-button">post</button>
-        <div class="dropdown-content" id="dropdown-content" [style.display]="isDropdownOpen ? 'block' : 'none'">
-            <form (ngSubmit)="submitForm()">
-                <input class="form-control" type="text" id="title" placeholder="title" [(ngModel)]="formData.title" name="title" required><input class="form-control" type="file" id="picture" [(ngModel)]="formData.picture" name="picture" accept="image/*" >
-
-                <textarea class="form-control" id="description" name="description" rows="4" placeholder="Whats in you heart" [(ngModel)]="formData.description" required></textarea>
-
-                <button type="submit" class="btn btn-success" id="post_submit">Submit</button>
-            </form>
-        </div>
-    </div>
-</div>
-<br><br><br>
-<!-- scrollable-and showing posts on screen -->
-<div class="data-container" (scroll)="onScroll()">
-    <div *ngFor="let item of data" class="data-item">
-        <div class="container text-center">
-            <div class="row justify-content-center">
-                <div class="col-6">
-                    <div class="card">
-                        <img class="card-img-top" src="{{picture}}" alt="User Picture">
-                        <div class="card-body">
-                            <h5 class="card-title">User ID: {{ userId }}</h5>
-                            <p class="card-text">{{ description }}</p>
-                            <button class="btn btn-primary">Discussion Box</button>
-                        </div>
-                    </div>
-                </div>
-            </div>
-        </div>
-    </div>
-</div>
-
